+++ src/Methods/OpenSsl.php
@@ -12,6 +12,7 @@ defined( 'ABSPATH' ) || exit;
 
 use CryptForWordPress\Helper;
 use CryptForWordPress\Method_Base;
+use RuntimeException;
 
 /**
  * Object to handle crypt tasks with OpenSSL.
@@ -24,6 +25,18 @@ class OpenSsl extends Method_Base {
 	 * @var string
 	 */
 	protected string $name = 'openssl';
+
+	/**
+	 * The method configurations.
+	 *
+	 * @var array<string,mixed>
+	 */
+	protected array $configuration = array(
+		'hash_type'        => 'hash',
+		'hash_algorithm'   => 'argon2',
+		'force_mu_plugin'  => false,
+		'cipher_algorithm' => 'AES-256-GCM',
+	);
 
 	/**
 	 * Instance of this object.
@@ -45,6 +58,9 @@ class OpenSsl extends Method_Base {
 
 	/**
 	 * Constructor for this object.
+	 *
+	 * @return void
+	 * @throws RuntimeException If error occurred.
 	 */
 	public function init(): void {
 		$this->set_hash( $this->get_hash_value_from_constant() );
@@ -59,7 +75,34 @@ class OpenSsl extends Method_Base {
 
 		// if no hash is set, create one.
 		if ( empty( $this->get_hash() ) ) {
-			$hash = hash( 'sha256', (string) wp_rand() );
+			// bail if configured hash algorithm does not exist.
+			if ( ! in_array( $this->configuration['hash_algorithm'], hash_algos(), true ) ) {
+				throw new RuntimeException( 'Unknown hash algorithm: ' . wp_kses_post( $this->configuration['hash_algorithm'] ) );
+			}
+
+			// prepare the hash.
+			$hash = '';
+
+			// use hash() to generate the hash.
+			if ( 'hash' === $this->configuration['hash_type'] ) {
+				$hash = hash( $this->configuration['hash_algorithm'], (string) wp_rand() );
+			}
+
+			// use hash_pbkdf2() to generate the hash.
+			if ( 'hash_pbkdf2' === $this->configuration['hash_type'] ) {
+				$hash = base64_encode(
+					hash_pbkdf2(
+						$this->configuration['hash_algorithm'],
+						(string) wp_rand(),
+						wp_salt(),
+						150000,
+						32,
+						true
+					)
+				);
+			}
+
+			// set resulting hash.
 			$this->set_hash( $hash );
 		}
 
@@ -72,7 +115,7 @@ class OpenSsl extends Method_Base {
 		}
 
 		// bail if wp-config.php is not writable.
-		if ( ! Helper::is_writable( $wp_config_php_path ) ) {
+		if ( $this->configuration['force_mu_plugin'] || ! Helper::is_writable( $wp_config_php_path ) ) {
 			$this->create_mu_plugin();
 			return;
 		}
@@ -126,6 +169,7 @@ class OpenSsl extends Method_Base {
 	 * @param string $plain_text Text to encrypt.
 	 *
 	 * @return string
+	 * @throws RuntimeException If error occurred.
 	 */
 	public function encrypt( string $plain_text ): string {
 		// bail if it is unusable.
@@ -138,13 +182,23 @@ class OpenSsl extends Method_Base {
 			return '';
 		}
 
-		$cipher    = 'AES-128-CBC';
+		// get the cipher algorithm.
+		$cipher = $this->configuration['cipher_algorithm'];
+
+		// bail if the configured cipher is not available.
+		if ( ! in_array( strtolower( $cipher ), openssl_get_cipher_methods(), true ) ) {
+			throw new RuntimeException( 'Unknown cipher algorithm: ' . wp_kses_post( $this->configuration['cipher_algorithm'] ) );
+		}
+
+		// gets the cipher iv length.
 		$iv_length = openssl_cipher_iv_length( $cipher );
 
+		// bail if iv length could not be loaded.
 		if ( ! is_int( $iv_length ) ) { // @phpstan-ignore function.alreadyNarrowedType
 			return '';
 		}
 
+		// get the iv.
 		$iv = openssl_random_pseudo_bytes( $iv_length );
 
 		// bail if iv could not be created.
@@ -152,14 +206,55 @@ class OpenSsl extends Method_Base {
 			return '';
 		}
 
-		$ciphertext_raw = openssl_encrypt( $plain_text, $cipher, $this->get_hash(), OPENSSL_RAW_DATA, $iv );
+		// handle GCM-based ciphers.
+		if ( $this->should_use_aead_tag( $cipher ) ) {
+			// set the tag.
+			$tag = '';
 
-		if ( ! $ciphertext_raw ) {
-			return '';
+			// encrypt the string.
+			$ciphertext = openssl_encrypt(
+				$plain_text,
+				$cipher,
+				$this->get_hash(),
+				OPENSSL_RAW_DATA,
+				$iv,
+				$tag
+			);
+
+            // bail if text could not be encrypted.
+            if( ! is_string( $ciphertext ) ) {
+                return '';
+            }
+
+			// return the resulting encrypted string.
+			return base64_encode(
+				base64_encode( $iv ) . ':' .
+				base64_encode( $tag ) . ':' . // @phpstan-ignore argument.type
+				base64_encode( $ciphertext )
+			);
+		} else {
+			// for all others ciphers.
+
+			// encrypt the string.
+			$ciphertext_raw = openssl_encrypt(
+				$plain_text,
+				$cipher,
+				$this->get_hash(),
+				OPENSSL_RAW_DATA,
+				$iv
+			);
+
+			// bail if anything failed.
+			if ( ! $ciphertext_raw ) {
+				return '';
+			}
+
+			// get the HMAC.
+			$hmac = hash_hmac( $this->configuration['hash_algorithm'], $ciphertext_raw, $this->get_hash(), true );
+
+			// return the resulting encrypted string.
+			return base64_encode( base64_encode( $iv ) . ':' . base64_encode( $hmac . $ciphertext_raw ) );
 		}
-
-		$hmac = hash_hmac( 'sha256', $ciphertext_raw, $this->get_hash(), true );
-		return base64_encode( base64_encode( $iv ) . ':' . base64_encode( $hmac . $ciphertext_raw ) );
 	}
 
 	/**
@@ -168,6 +263,7 @@ class OpenSsl extends Method_Base {
 	 * @param string $encrypted_text The encrypted string.
 	 *
 	 * @return string
+	 * @throws RuntimeException If cipher is unknown.
 	 */
 	public function decrypt( string $encrypted_text ): string {
 		// bail if it is unusable.
@@ -180,36 +276,111 @@ class OpenSsl extends Method_Base {
 			return '';
 		}
 
-		$cipher    = 'AES-128-CBC';
+		// get the cipher algorithm.
+		$cipher = $this->configuration['cipher_algorithm'];
+
+		// bail if the configured cipher is not available.
+		if ( ! in_array( strtolower( $cipher ), openssl_get_cipher_methods(), true ) ) {
+			throw new RuntimeException( 'Unknown cipher algorithm: ' . wp_kses_post( $this->configuration['cipher_algorithm'] ) );
+		}
+
+		// gets the cipher iv length.
 		$iv_length = openssl_cipher_iv_length( $cipher );
+
+		// bail if iv length could not be loaded.
 		if ( ! $iv_length ) {
 			return '';
 		}
+
+		// decode the encrypted text.
 		$c = base64_decode( $encrypted_text );
-		if ( str_contains( $c, ':' ) ) {
+
+		// handle GCM-based ciphers.
+		if ( $this->should_use_aead_tag( $cipher ) ) {
+			// get the parts.
 			$c_exploded = explode( ':', $c );
+
+			// get the part contents.
 			$iv         = base64_decode( $c_exploded[0] );
-			if ( ! $iv ) {
-				return '';
-			}
-			$iv = substr( $iv, 0, $iv_length );
-			$c  = base64_decode( $c_exploded[1] );
-			if ( ! $c ) {
-				return '';
-			}
-			$hmac           = substr( $c, 0, $sha2len = 32 );
-			$ciphertext_raw = substr( $c, $sha2len, strlen( $c ) );
+			$tag        = base64_decode( $c_exploded[1] );
+			$ciphertext = base64_decode( $c_exploded[2] );
+
+			// decrypt the text.
+			$original_plaintext = openssl_decrypt(
+				$ciphertext,
+				$cipher,
+				$this->get_hash(),
+				OPENSSL_RAW_DATA,
+				$iv,
+				$tag
+			);
+
+            // bail if decryption was not successfully.
+            if ( ! is_string( $original_plaintext ) ) {
+                return '';
+            }
 		} else {
-			$iv             = substr( $c, 0, $iv_length );
-			$hmac           = substr( $c, $iv_length, $sha2len = 32 );
-			$ciphertext_raw = substr( $c, $iv_length + $sha2len );
+			// for all other ciphers.
+
+			// backwards-compatibility for strings that does not contain ":".
+			if ( str_contains( $c, ':' ) ) {
+				// get the parts.
+				$c_exploded = explode( ':', $c );
+
+				// get IV.
+				$iv = base64_decode( $c_exploded[0] );
+
+				// bail if IV could not be loaded.
+				if ( ! $iv ) {
+					return '';
+				}
+
+				// get IV part.
+				$iv = substr( $iv, 0, $iv_length );
+
+				// get HMAC part.
+				$c = base64_decode( $c_exploded[1] );
+
+				// bail if HMAC part could not be loaded.
+				if ( ! $c ) {
+					return '';
+				}
+
+				// get HMAC.
+				$hmac = substr( $c, 0, $sha2len = 32 );
+
+				// get the raw ciper text.
+				$ciphertext_raw = substr( $c, $sha2len, strlen( $c ) );
+			} else {
+				$iv             = substr( $c, 0, $iv_length );
+				$hmac           = substr( $c, $iv_length, $sha2len = 32 );
+				$ciphertext_raw = substr( $c, $iv_length + $sha2len );
+			}
+
+			// decrypt the string.
+			$original_plaintext = openssl_decrypt( $ciphertext_raw, $cipher, $this->get_hash(), OPENSSL_RAW_DATA, $iv );
+
+			// bail if decryption was not successfully.
+			if ( ! is_string( $original_plaintext ) ) {
+				return '';
+			}
+
+			// get the hashed HMAC.
+			$calc_mac = hash_hmac( $this->configuration['hash_algorithm'], $ciphertext_raw, $this->get_hash(), true );
+
+			// bail if HMAC is not set.
+			if ( empty( $hmac ) ) {
+				return '';
+			}
+
+			// bail if HMAC does not match.
+			if ( ! hash_equals( $hmac, $calc_mac ) ) {
+				return '';
+			}
 		}
-		$original_plaintext = openssl_decrypt( $ciphertext_raw, $cipher, $this->get_hash(), OPENSSL_RAW_DATA, $iv );
-		$calc_mac           = hash_hmac( 'sha256', $ciphertext_raw, $this->get_hash(), true );
-		if ( $original_plaintext && $hmac && hash_equals( $hmac, $calc_mac ) ) {
-			return $original_plaintext;
-		}
-		return '';
+
+		// return the resulting decrypted string.
+		return $original_plaintext;
 	}
 
 	/**
@@ -224,6 +395,17 @@ class OpenSsl extends Method_Base {
 		// save the hash in the database.
 		update_option( $this->get_slug() . '_hash', $this->get_hash() );
 
+		// run parent uninstalling tasks.
 		parent::uninstall();
+	}
+
+	/**
+	 * Return whether the given cipher should use AEAD with tag.
+	 *
+	 * @param string $cipher The cipher name.
+	 * @return bool
+	 */
+	private function should_use_aead_tag( string $cipher ): bool {
+		return str_contains( strtolower( $cipher ), 'gcm' ) || str_contains( strtolower( $cipher ), 'poly1305' );
 	}
 }