+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.2] - 15.03.2026
+
+### Added
+
+- Added configurations for OpenSSL and Sodium
+- Added configuration for force one method by setting instead of PHP hook
+
 ## [1.0.0] - 14.03.2026
 
 ### Added